[RATOM-113] Restructures tests.
