+++ api/tests/conftest.py
@@ -1,5 +1,4 @@
 import pytest
-from unittest import mock
 from rest_framework.test import APIClient
 
 
@@ -12,9 +11,3 @@ def api_client_anon():
 def api_client(api_client_anon, user):
     api_client_anon.force_authenticate(user=user)
     yield api_client_anon
-
-
-@pytest.fixture
-def celery_mock():
-    with mock.patch("api.views.account.import_file_task") as _mock:
-        yield _mock
